Fixing seg fault in code-generating structural types
With Steve's revision to reconfigure how reference counting is done
for domains and arrays, my code to generate type/offset information of
communicated buffers started to seg fault.  With a bit of
investigation it turned out that the reason was that some of the types
that I was registering as needing their structural interfaces
generated were culled from the AST prior to codegen making my pointer
stale.

Originally, I thought that the most reasonable thing to do would be to
register these types when they were created -- in the case of
chpl_comm_fork, that would be in the parallel pass.  However, this
approach didn't work for communication (or was too hard to do as an
early pass), so for the macro wrappers around chpl_comm_gets/puts, I
was doing the registration at codegen time.  It turns out that doing
the chpl_comm_fork calls early resulted in this dead AST problem, so I
am now handling them at codegen time as well which fixes the problem.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15741 3a8e244f-b0f2-452b-bcba-4c88e055c3ca