Fixing a bug with the type/offset code generation mode
Lee found a bug in my code yesterday where I generated an
extra underscore in the invocation of the SPECIFY_SIZE()
macro for chpl_comm_fork/chpl_comm_fork_nb calls if the
genCommunicatedStructures flag was off.  At the time that
I implemented this, it was yet another case of needing
an extra underscore when I was generating chpl_rt_type_id
enumerations for reasons that I couldn't explain.
Unfortunately, I was equating the genCommunicatedStructures
flag with CHPL_COMM_HETEROGENEOUS being on in the runtime
when in fact there isn't necessarily any causal relationship
between the two.  The effect of this was that when Lee was
debugging the PVM comm layer (which turns genCommunicatedStructures
on) with CHPL_COMM_HETEROGENEOUS off, we were generating
sizeof() expressions that evaluated to the pointer type's size
(8 bytes) rather than the C struct that the pointer was
referring to, causing insufficient data to be sent to the
chpl_comm_fork calls.  I was also seeing seg faults due to
this when debugging the genCommunicatedStructures flag under 
GASNet with CHPL_COMM_HETEROGENEOUS turned off, but never had
the chance to dig into it and see what was wrong.  Sorry Lee.

This checkin fixes this problem by (a) removing the conditional
based on genCommunicatedStructures, and (b) using a saner
heuristic for determining when to generate the extra underscore
-- in particular, we only generate it now when code generating
a CLASS_CLASS that does not have FLAG_REF set (is not a reference
type).

Since this change permitted me to test compiling GASNet with
genCommunicatedStructures on, it also allowed me to uncomment
the reference to chpl_rt_type_id in the non-HETEROGENEOUS
case of chplcomm.h and test the release/examples directory
under GASNet and for the first time all of them compile and
run correctly in this mode -- a sign that this is getting
closer to being correct (though I'm still not confident that
all the information in the HETEROGENEOUS case is set up
correctly, so more debugging work to do there).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15742 3a8e244f-b0f2-452b-bcba-4c88e055c3ca