Added developer flag for generating type/offset information for communication
To aid with debugging, I added a developer flag
--[no-]gen-communicated-structures that turns the
generation of the _type_structure.c file on and
off.  By default, it is set to on if CHPL_COMM is
PVM and off otherwise.  We didn't look at driver.cpp
in yesterday's code review due to time constraints,
but one other TODO item that's on my list is that
remote value forwarding currently breaks the
generation of the type/offset information, so I
turn remote value forwarding off if 
--gen-communicated-structures is on.  This code
moved with the addition of the flag.

I also investigated one of the mysterious underscore
cases from the code review yesterday and found it to
be unnecessary, so removed it.  I think it must have
stemmed from a point when I didn't have all the types
flowing through the CHPL_COMM macros appropriately.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15746 3a8e244f-b0f2-452b-bcba-4c88e055c3ca