This change aligns the launcher to the order of the architectures in args.c.
This change aligns the PVM launcher to the order of the architectures in
args.c. This resulted in hello-multirealm using the wrong architecture in
its output. Also, this change fixes a bug when adding nodes in a realm
number greater than another's whose total locales is 0 (resulting in a
segfault).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15749 3a8e244f-b0f2-452b-bcba-4c88e055c3ca