This change allows the PVM launcher to use different files on different archs.
This change allows the PVM launcher to use different files on different
architectures. The current heuristic is to capture the value in the
environment variable CHPL_MULTIREALM_LAUNCH_DIR_xxxxx, where xxxxx is the 
target architecture. If this isn't set, then the default is CHPL_HOME. We
also capture the application via argv[0] (delimited by '/'). Combine these
two values (in that order) and append '_real.' Next, anywhere the current
architecture is used in the application name, replace that with the
target architecture (hello_darwin may get replaced with hello_linux). Also,
I've pulled a lot of the freeing where the next immediate step is some
form of exit. I'll rework that eventually, but it was confusing valgrind,
and it sometimes led to glibc errors.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15752 3a8e244f-b0f2-452b-bcba-4c88e055c3ca