This change updates the PVM launcher so that launched files can be arbitrary.
This change updates the PVM launcher such that the shared directory
requirement is no longer necessary. Prior to this change, the given file
would be tested prior to execution; if the file were not present (or not
readable), the launcher would tell you to make sure the given file existed.
However, if this file did exist on a local filesystem, the message would
be false, and the program would exit. Now, the launcher does not check.
It builds the command from the CHPL_MULTIREALM_LAUNCH_DIR_<realm>
environment variable (which would be the local CHPL_HOME if not set locally).
If $CHPL_MULTIREALM_LAUNCH_DIR_<realm>/app_real can't be spawned (PvmNoFile),
it tries with whatever was given on the command line. If this fails, then
we can safely error out. Debugging information is left in for now, though
the lines are commented out.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15753 3a8e244f-b0f2-452b-bcba-4c88e055c3ca