This fix allows the setup.sh utility to propegate CHPL_COMM and CHPL_DEVELOPER.
This fix allows the setup.sh utility to propegate the environments CHPL_COMM
and CHPL_DEVELOPER to the remote nodes for which the user wants to compile.
Prior to this fix, CHPL_COMM=none was used (and CHPL_DEVELOPER was not set).
A later addition to this setup.sh will change how the command line works.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15755 3a8e244f-b0f2-452b-bcba-4c88e055c3ca