This change alters the way command-line arguments are proccessed in setup.sh.
This change alters the way command-line arguments are proccessed in setup.sh.
If setup.sh is run without arguments, it looks for $CHPL_HOME/util/CONFIG.
Otherwise (there are arguments), if the first character of the first argument
is a '-,' it assumes it's a flag for chpl, and everything that follows is a
flag. Otherwise, if the first argument is a readable file and that file
contains the string HOST in it, it assumes that is the configuration file, and
everything else becomes the flags. Otherwise, all arguments are treated as
flags. Note that if the arguments passed are .chpl files to compile, the
setup.sh still doesn't really process that right. For example, if the user
runs setup.sh -o hello hello.chpl, and there aren't any files in TEST in the
configuration file $CHPL_HOME/util/CONFIG, nothing will be compiled. If there
are several Chapel files listed in TEST, then what's on the command line will
be compiled each time for each system (going into the same directory w/o
changing names).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15756 3a8e244f-b0f2-452b-bcba-4c88e055c3ca