This change adds a piece to the PVM launcher in case _real file isn't found.
This change adds another layer to the PVM launcher when spawning. Before, if
the xxxx_real file wasn't found in either CHPL_MULTIREALM_LAUNCH_DIR_<realm>
or in what was passed into argv[0], then the launcher would fail. Using the
start_test script in util changes the directory frequently, and it does not
expland the binary name, so this change adds $PWD. (Note 'cwd' doesn't work
if passed to getenv.)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15759 3a8e244f-b0f2-452b-bcba-4c88e055c3ca