This change adds PVM support for put/get/fork counters.
This change adds PVM support for put/get/fork counters in the same way used
by GASNET. Also, it adds a couple cases to the table for heterogeneous
execution. First, it checks ahead of time if the thing being sent is a (nil).
If so, it doesn't do anything (instead of potentially dereferencing 0x0).
This commit also adds chpl_bool and chpl_string, the latter of which can't
work until the sizeof the string can be determined.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15762 3a8e244f-b0f2-452b-bcba-4c88e055c3ca