This commit ensures the PVM comm layer sends are synchronous.
This commit ensures the PVM comm layer sends are synchronous. There is no
synchronous PVM mode (ala MPI_Ssend), so to simulate that, after every
pvm_send (of real data -- "metadata" cases are always followed by a
corresponding action), send an ack back to the sender. Unfortunately, PVM
requires packages to be sent with actual data, so we're just sending a 0.
Should we want to do so later, we could send status, but for now, this is
enough to ensure synchronous completion. Without this change, often variables
would get used on multiple locales without being updated leading to
incorrect and non-deterministic results.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15765 3a8e244f-b0f2-452b-bcba-4c88e055c3ca