This change alters how PVM comm layer code starts and finishes.
This change does two things. First, it checks to see if a PVM daemon is
already running, and if so, it does not attempt to start a new one. (In
doing this, the launcher creates a temporary file, /tmp/chpl.pvmlaunch.txt,
and it stuffs first the user ID and then the name of the PVM daemon temp
file, if it exists.) Second, when the comm layer is tearing down, this
change allows every node to exit cleanly. Instead of using pvm_halt, which
prints a 'Terminated' message, it uses pvm_exit. This will leave PVM
running, but since each node has reported back to the parent (launcher), all
nodes will have finished, and the pvm_delhosts call will make sure only the
laucher node stays configured.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15770 3a8e244f-b0f2-452b-bcba-4c88e055c3ca