This change fixes chpl_comm_exit_any in the PVM comm layer.
This change addresses error conditions in chpl_comm_exit_any for the PVM
comm layer. Prior to this fix, whenever a Chapel application compiled for the
PVM layer exited on an error (for example, configBadValue reported an error
on a type and exited), PVM would get into a bad state, and applications
would be left hanging. The change mimics what exit_all does, so now,
applications exit cleanly, allowing further applications to run.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15772 3a8e244f-b0f2-452b-bcba-4c88e055c3ca