Added a set of functions called chpl__autoDestroy that the compiler can insert to destroy value type variables.  This is currently used only when destroying global variables, but will be useful for local variables as well, eventually.
Added a compiler-inserted function chpl__autoDestroyGlobals that
contains calls to chpl__autoDestroy on all global variables, in order
to destroy and free all global variables.

Disabled and removed code to free local variables since this relies on
optimizations and conflicts with freeing global variables.  Simplified
the prune pass so that it does not need to hold onto the destructors
on its first call.  Futurized 8 tests that explicitly checked for
compiler-inserted destructor calls and frees of local memory.

Removed initialize method on array and domain wrapper records.  This
method incremented the reference count.  Rather than increment, this
is now simply set to one in the constructor.

Added "call destructor" primitive that is used in the
chpl__autoDestroy functions to call the destructors of records.  This
is different than calling 'delete' since we don't need to free
records.

Added "no auto destroy" pragma that can be applied to variables (local
and global, though it only matters for global now) in order that they
not be destroyed by the compiler.  Applied this pragma to memLeaksLog
which is examined after globals are destroyed.

This change also includes revision 15743 which I tried earlier this
month and then backed out in revision 15748.  The relevant part of the
log message for this commit reads as follows:

  "Reworked the cullOverReferences pass so that it no longer removes
   all references to arrays and domains but rather it removes the
   reference from array and domain references that are returned by
   functions.  This is still consistent with commit 15730 but it is
   more specific.  In particular, we will need references to arrays
   and domains to deal with other work that I am doing to allow us to
   free local arrays and domains.

   Modified transform that allocates locals on the heap to expect
   references where there previously were none.

   Modified optimization that removes array and domain records to
   expect references where there previously were none."


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15776 3a8e244f-b0f2-452b-bcba-4c88e055c3ca