Adding some user-defined reduction tests based on conversation with Bin Ren
Bin Ren tried typing in the user-defined reduction from our old paper
which didn't work primarily due to a number of naming changes (but
also because the constructor-style syntax we suggested there has never
been implemented).  This captures that code and some attempts to get
it working, most of which didn't work and are filed here as bug
futures:

* mink-paper.chpl: the original version

* mink-minchange.chpl: the minimal set of changes that are required to
  get it mostly working; this doesn't quite work because the first
  array element gets corrupted because of a premature free (thanks to
  David for the diagnosis)

* mink-blc.chpl: a slight cleanup of the previous version that I
  implemented; doesn't work for the same reason

* mink-blc-infer.chpl: a version that infers the return type of the
  reduction which gets a nil dereference at runtime.  This seems
  related to the same issue because of the following.

* mink-blc-workaround.chpl: a workaround suggested by David in which
  we insert an explicit temp into the generate method.

Thanks for the detective-work David!




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15788 3a8e244f-b0f2-452b-bcba-4c88e055c3ca