Comment out the delete call that frees reduction classes.
Since the destructors call the destructors for any array member variables in
the reduction class, arrays returned as the result of a reduction were getting
corrupted.

Once array reference counting is working well, the line that inserts the
destructor call in compiler/AST/build.cpp can be uncommented.

Added a future that will fail until the delete call is being inserted again.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15797 3a8e244f-b0f2-452b-bcba-4c88e055c3ca