Added "no copy" pragma for use in internal module code where we do not want to insert an _copy function to copy an initialization expression in an inferred variable declaration.
Applied "no copy" pragma to iterator class instances in order to avoid
creating arrays while capturing an iterator class into a variable that
we can explicitly free.

This change improves memory management of iterator classes ("iterator
data") over a run of the test suite (from 55.9KB to 8.3KB).

At least some of the remaining leak is an interesting case.  If the
compiler inserts code to free data within an iterator, we may not call
it if we only run a single iteration of the iterator or, in general,
if we break out of the loop before the iterator has run to completion.
To fix this, we'll need to have the compiler create code to free data
in the case that an iterator is not run to completion.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15799 3a8e244f-b0f2-452b-bcba-4c88e055c3ca