Fixed a bug in which a 'copied' tuple could alias its components.  For example, in the statement 'var t = (A, B)' where A and B are arrays, the components of the tuple t(1) and t(2) would be aliases.  Contrast this with 'var AA = A' in which AA would be a copy of A.  This checkin fixes this bug.  Updated test_strings_and_tuples test case to reflect this semantic change.
Renamed the internal function "_copy" to "chpl__initCopy" which more
accurately reflects its role of copying the values of initialization
expressions.  Reimplemented this function for tuples to copy each of
the components.  Moved all such function definitions to the same place
in the ChapelBase module.

Fixed two test cases in which compiler error line numbers changed.
The line numbers reported in nilDynamicDispatch were wrong and are
still wrong.  This should be examined at some point.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15802 3a8e244f-b0f2-452b-bcba-4c88e055c3ca