This change to the PVM launcher fixes testVFlag.chpl and testVerboseFlag.chpl.
This change to the PVM launcher does three things. First: testVFlag.chpl and
testVerboseFlag.chpl was printing an extra newline. By nature of -v, the
command gets printed out; the output is also broken in all other comm layers.
In other comm layers, the return from the chpl_launch_using_system has to
be the command (so that's what gets spawned). But, in the PVM layer, all the
work is done prior to hitting the return, so the return was a "". By exiting
prior to the return, it fixed the problem thereby matching the output.
Second: added signal handling to the launcher, though it doesn't work yet.
Third: cleaned up some of the code. pvm_setopt can quiet libpvm output, so
I don't need to build auxiliary lists of nodes to pass to pvm_add/delhosts
so duplicate nodes aren't used (thus causing an error).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15805 3a8e244f-b0f2-452b-bcba-4c88e055c3ca