Make labels on statements other than for, while-do and do-while loops be errors, make break and continue only be applicable in the those loop types and remove goto from the scanner and parser.
To do this, I added fields to the BlockStmt class for the break-label,
continue-label and user-label. They are NULL unless the block statement is
a loop that can have break and continue statements. The user label is used
to choose which break-label or continue-label to use as the target of a
labeled break or continue, but no label is actually generated in C for it.

Fixed a case when checking control flow of parallel statements that wasn't
allowing for example: "begin { for i in 1..n do if i == c then break; }"
because it expected non-labeled breaks to have a NULL label instead of it
being the gNil symbol.

Updated the "Label, Break and Continue" spec subsection to be more specific
about what is and isn't allowed, and added an example.

Removed tests that use labels on non-loop statements and use goto in
uninteresting ways.

Refactored classes/deitz/infer/infer_field2*.chpl by putting their common code
into a new file and using that module from each test. These test are simulating
the transformation the compiler does for iterators. In these tests and in
test_iterator_transform.chpl and test_tech_forum_071004.chpl, I replaced a few
goto statements with conditionals. I left the old versions in commented out
sections.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15811 3a8e244f-b0f2-452b-bcba-4c88e055c3ca