This change cleans up memory utilization in the PVM launcher.
This change cleans up memory utilization in the PVM launcher. There are now
10 flags representing the 10 variables that get chpl_malloc'ed over the course
of the lifetime of the launcher. On exit, a function called memory_cleanup()
checks the defined mask to see which variables have been allocated (and
chpl_frees them). Variables are cleaned up as soon as they are no longer
used.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15820 3a8e244f-b0f2-452b-bcba-4c88e055c3ca