Add more functionality to the block distribution
Added member, order and position methods to the block domain class. They each
just call the function with the same name on the domain's whole field.

Forward the stridable parameter to the non-distributed domains used to
represent the whole array and blocks of the array.

Added array and domain slice functions to the block distribution. The array
slice function creates an alias of each of the local arrays on the locale
they are on.

Added some tests of these features.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15824 3a8e244f-b0f2-452b-bcba-4c88e055c3ca