Quick fix to re-enable OPTIMIZE=1 builds of compiler
When building using OPTIMIZE=1 as a developer (as our test system does),
we got a "variable may not be initialized before used" for a variable
'rhs' in Albert's new code.  I initialized it to NULL for now, but it 
looks to me like it is correct that the variable may not be initialized
if neither branch of the conditional is taken -- should an else clause
with an INT_FATAL be added?  I wasn't sure and didn't have time to reason
through it, so just made the error go away for now without making things
any worse.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15833 3a8e244f-b0f2-452b-bcba-4c88e055c3ca