Change range.sharedInternalIterator() into the default serial range iterator.
It had a comment explaining that it was a workaround for a circularity problem
in function resolution. Resolving the serial iterator caused the parallel
iterator to be resolved which depended on the type of the serial iterator.

This function was preventing zipping Block domains with ranges because the
Block domain leader iterator yields domains instead of ranges, and the range
follower was calling sharedInternalIterator on them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15839 3a8e244f-b0f2-452b-bcba-4c88e055c3ca