(Paired with Lee)
Added compiler support for a 'private broadcast table', called
chpl_private_broadcast_table, in the generated code that stores the
addresses of all symbols in the generated code as well as a few
symbols in the runtime.  Changed the interface of the private
broadcast function chpl_comm_broadcast_private to take an index into
this table rather than an address.  This will allow (eventually) the
runtime code to be implemented in such a way that we do not assume the
addresses of these symbols are the same on different locales.  This is
a poor assumption in general and immediately problematic in the case
of heterogeneous locales.  Note that the runtime implementation is
unchanged with this commit; the function indexes into the table
immediately and uses the address (as opposed to communicating the
index) so it makes the same assumption.

The runtime constants chpl_verbose_comm, chpl_comm_diagnostics, and
chpl_verbose_mem are always placed at the beginning of the private
broadcast table so that they can continue to be communicated using the
private broadcast function.

In the compiler, the table is generated at codegen time based on all
calls to the private broadcast function.

Unduplicated a swath of duplicated compiler code in the GPU work
related to the globals registry and the memory description strings
table.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15845 3a8e244f-b0f2-452b-bcba-4c88e055c3ca