Fixing a bug in TAB search script
While the TAB search script was set up to check out a copy of the
sources and set them, it actually looked in CHPL_HOME which was
never set by the script -- so it was searching my copy of the
sources (generally reasonably up-to-date, but still...).  This
check-in fixes the issue and sets CHPL_HOME before looking.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15847 3a8e244f-b0f2-452b-bcba-4c88e055c3ca