This change fixes PVM's broadcast_private and adds some 64->32-bit support.
This change changes the chpl_comm_broadcast_private() routine in the PVM comm
layer to adequately handle the new mechanism. If you are node n, and you
want to send a private broadcast variable to nodes m, o, and p, you first
generate a tag, then you set up a _chpl_message_info structure (overloading
the union variable fid to be the new id into chpl_private_broadcast_table),
and send the structure (with the TAGMASK+1 tag) to nodes m, o, and p. The
other nodes' polling threads will collect this, and upon reading the type
ChplCommBroadcastPrivate, will wait for another send. Your next send is of
chpl_private_broadcast_table[id] (which is an address of the thing you want
to update, say numLocales). The other change is to begin some of the work
into converting 64-bit things sent to 32-bit machines. This work is not
complete. Note: this change doesn't yet work for CHPL_COMM_HETEROGENEOUS.
The biggest flaw: I need to account for 'size' as a type.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15849 3a8e244f-b0f2-452b-bcba-4c88e055c3ca