Reworked how the compiler inserts code to put local variables on the heap.  This used to be done by assuming that the first def of the variable was an appropriate place to allocate the variable on the heap.  This required us to build up the def maps by traversing the IR in lexical order.  This assumption has always been worrisome and was marked by a comment containing the alarming phrase "ack!! this is troublesome".
This assumption is not valid with changes I am working on to free
memory associated with local variables.  In those changes, if a
function returns a record, we face the problem that the destructor of
that record may be called inside the function.  To deal with this
issue, we transform the code so that instead of returning the record,
we pass the record by reference to the function and copy it within.
However, if this applies to the first def of a variable, then making a
reference to that variable (a use instead of a def), invalidates this
assumption.

This change makes it so that local variables are put on the heap at
their defPoint (a DefExpr) rather than an actual def of the symbol.
This assumes that the defPoint is always before any use or def of the
variable.  This assumption seems much safer.

Changed the code that normalizes alias definitions so that the
defPoint is before the def, in order to make the new assumption valid.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15856 3a8e244f-b0f2-452b-bcba-4c88e055c3ca