This change allows CHPL_COMM_HETEROGENEOUS to work on private broadcasts.
This change allows CHPL_COMM_HETEROGENEOUS to work on private broadcasts.
When the AST is pulling out PRIM_PRIVATE_BROADCAST, it grabs the type of
the variable; without CHPL_COMM_HETEROGENEOUS, the sizeof(type) is
sufficient, but with it, this must be converted to the chpl_rt_type index.
SPECIFY_SIZE does this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15862 3a8e244f-b0f2-452b-bcba-4c88e055c3ca