Added a new primitive type, chpl_threadID_t, to support different pthread_t types between Linux and OS X. The underlying C type and its default value are defined in the header file for each threading layer.
In the compiler, the type supports equality comparison against itself
and casts to other primitive types. Casting to a 64-bit integer type
will generally be the safest cross-platform way to for thread IDs to
interact with other types.

The primitive "thread_id" has always called chpl_thread_id(); since
that this function now returns a chpl_threadID_t, the primitive now
also returns this type.

A default hash function for associative domains over chpl_threadID_t
was created; it casts to int(64) and uses that default hash function.

IO writes for the type is done by first casting to a uint(64) and then
writing that value.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15874 3a8e244f-b0f2-452b-bcba-4c88e055c3ca