Added code to the default arithmetic array destructor in order to call autoDestroy on all of the array's elements.  This is necessary to free the memory in arrays of arrays and arrays of domains, and to call the destructors for arrays of records, etc.  This has a big impact on the test users/jglewis/badGetModule in which used to leak 196MB of memory.  It now leaks 68B of memory, all strings.
Reimplemented the memory tracking hash table so that it doubles in size if it
ever gets over half full and halves in size if it ever gets less than an eighth
full.  Now that we are freeing memory, this hash table becomes very slow as
discussed in r15615.  This makes the test above tractable, even fast.  Without
this change, it would time out.

Also refactored chpl_init_mem_table since this can't really be initialized
until the config variables are parsed and broadcasted.  This means that we
don't track anything prior to this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15880 3a8e244f-b0f2-452b-bcba-4c88e055c3ca