This change fixes strings in the PVM layer with CHPL_COMM_HETEROGENEOUS on.
This change fixes strings in the PVM layer with CHPL_COMM_HETEROGENEOUS on.
The only time SPECIFY_STRING_SIZE is used within the comm layer is for the
function CHPL_COMM_WIDE_GET_STRING. If we trick the comm layer to take a
negative number for the size, we can overload the meaning of size to be
(instead of the index into the chpl_fieldType table) the chplType
enumerated value of interest. -CHPL_TYPE_chpl_string is -16 under the
current implementation. An analogous (and still to be checked in) change in
the comm layer sees the negative number, makes it positive, and sends via
pkstr. This works because pvm_pkstr reads until it sees a null.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15896 3a8e244f-b0f2-452b-bcba-4c88e055c3ca