Removting subBlocks iterator interface from distributions
Sung (and Steve?) asked about the status of the subBlocks
iterator in the distributions today.  Looking at it, it's
clear that in the face of leader/follower iteration, it
is not as crucial as it used to be when we only had task
parallelism and no forall loops.  Moreover, only a few
tests, mostly outdated, referenced it at all.  Removed
it from the distributions that implemented it, tests
that defined distributions that also implemented it,
and the one main test that still referred to it -- a copy
of RA (the one we use for performance timings, ironically,
giving me yet another reason to update our performance
graphs).  Deleted a second test (ra-paramWidth) that 
referred to it, but whose initial log message said that
the test could be retired once it was working and its
trick was worked into the mainline RA which it is and
has been for some time now.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15906 3a8e244f-b0f2-452b-bcba-4c88e055c3ca