This change adds a new table, chpl_sizeType, which broadcasts structure sizes.
This change adds a new table, chpl_sizeType, which broadcasts structure sizes.
The table is analogous to the chpl_rt_types enumeration, except it is an
array of sizeof(<whatever>). The nth enumerated element of the former
translates directly to the nth element of the array chpl_sizeType. A new
function, chpl_getFieldSize, simply returns the sizeof the nth chpl_rt_types
element.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15907 3a8e244f-b0f2-452b-bcba-4c88e055c3ca