This change incorporates compiler changes into PVM comm layer, fixes memory.
This change makes a few changes to the PVM comm layer. First, memory
allocation was incorrect when CHPL_COMM_HETEROGENEOUS was on. When a
chpl_comm_fork(_nb) was called, the size allocated for the RPC arguments
was represented by size; given that 'size' was overloaded to mean the
entry into the structType table, we needed to first convert back to an
actual size (using the new sizeType table). Next, a quick fix was added in
that chpl_LocaleID's maximal value is chpl_numLocales - 1. (This broke
receive and send code in very rare conditions.) Third, when the receive and
send code receives a negative size, it means that the send or receive came
directly from the comm layer (or chplcomm.h) as opposed to the compiler, so
instead of converting the structType to a rtType, we simply negate the
negative and assume it is an rtType. (The two places a negative size are now
sent is on broadcast_private of chpl_verbose_comm, chpl_comm_diagnostics,
and chpl_verbose_mem, and in chplcomm.h when CHPL_COMM_GET_WIDE_STRING is
invoked.) Finally, made a quick debug change whereby the globals broadcast
printf tells you directly what was sent (as opposed to the address of where
the regustry entry lives). A note on the private broadcast of
chpl_verbose_comm, chpl_comm_diagnostics, and chpl_verbose_mem: a new macro
is defined to handle this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15910 3a8e244f-b0f2-452b-bcba-4c88e055c3ca