Teased the single modules/standard directory into standard/ and internal/
Broke our modules/standard directory into two -- one containing
internal modules -- ones that the user shouldn't really need to
know about; and standard modules -- ones that are more like
standard libraries, are documented in the spec, etc.

As an unintentional result of my previous check-in, standard
modules used by internal modules were incorrectly parsed and
tagged as MOD_INTERNAL.  This change revealed that issue and
fixed it by postponing the parsing of these modules until
after the internal modules have been read in.

Improved our facility for printing out filenames of Chapel
sourcefiles so that rather than looking for "modules/standard"
and using that as a cue to replace its prefix with $CHPL_HOME,
actually looks for the value of $CHPL_HOME and replaces it
with the string "$CHPL_HOME" if it finds it.

Updated two tests that reported errors on modules/standard
modules that are now modules/internal modules.  Now that we
can distinguish between these two cases, this raises the
question of whether we want to report any errors/line#s for
modules/internal modules (for non-developers) or instead just 
say something like <built-in>.

Updated another test that prints out all non-internal files
parsed to reflect the standard files that we always parse
due to dependences in the internal modules.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15915 3a8e244f-b0f2-452b-bcba-4c88e055c3ca