Test of mixing up user module with standard
This test demonstrates that if a user has a file whose name
matches a standard module, it could cause problems whether
they use it (in foo2.chpl) or not (as in foo.chpl).  I have
a plan on how to deal with this -- just trying to get back
to more of a test-first philosophy before implementing it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15916 3a8e244f-b0f2-452b-bcba-4c88e055c3ca