Rejiggered how the compiler handles autoCopy and autoDestroy functions for calls that are changed from reference functions to value functions.  (Recall that the compiler changes a reference function into a value function in the case where the returned value of a 'var' function is not used as an lvalue.)  Prior to this change, when changing from a reference function to a value function, the compiler would add the "insert auto copy" and "insert auto destroy" flags to the new temp that captures the returned value from the value function if the type of the temp had an autoCopy function defined for it.  This works but is too conservative (too many autoCopy and autoDestroy calls).  In code that I am working on to remove autoCopy and autoDestroy on getters, this conservative code ends up reinserting such autoCopy and autoDestroy calls.
Originally I tried to change this code so that we only add these flags
to the new temp that captures the result of the value function if the
temp that captures the result of the reference function had these
flags.  However, this doesn't work because in some cases we need
autoCopy on the value function even though we didn't need it on the
reference version.  This is (arguably) because references are not
autoCopied but values are.

This commit makes it so that we use the original conditional (a call
to requiresImplicitDestroy) on the new call of the value function
(changed from the reference function) to make this decision.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15920 3a8e244f-b0f2-452b-bcba-4c88e055c3ca