Added a flag "no implicit copy" that applies to functions and keeps the compiler from inserting an autoCopy and autoDestroy call on a temporary inserted to capture the result of functions to which this flag applies.  Applied this flag to compiler-inserted getter/setter methods of record/class/union fields.
This commit has a major impact on the performance of certain tests.
In particular, tests of arrays and domains with distributions that are
implemented via classes which contain array and/or domain fields.
Prior to this change, every access of such a field would result in
(unnecessary) reference counting of the array and domain.  For
example, the execution time of test/npb/cg/bradc/cg-sparse.chpl (which
recently started timing out) when compiled with --no-local and
--no-checks improves greatly with this commit:

  Before this commit   7m45.153s on chevy-lnx
  After this commit    0m27.274s on chevy-lnx

However, before revision 15875 (initial support for inserting calls to
the destructors of records via autoCopy and autoDestroy), the time was
even better: 0m10.473s on chevy-lnx.  So there is more optimization
work to do, some of which I'm ready to test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15921 3a8e244f-b0f2-452b-bcba-4c88e055c3ca